+++ packages/terra-button/src/index.js
@@ -1,9 +1,10 @@
 import Button from './Button';
 import IconButton from './variants/_IconButton';
+import AnchorButton from './variants/_AnchorButton';
 import { ButtonTypes, ButtonVariants, IconTypes } from './_constants';
 
 export default Button;
 
 export {
-  IconButton, ButtonTypes, ButtonVariants, IconTypes,
+  IconButton, ButtonTypes, ButtonVariants, IconTypes, AnchorButton,
 };